A Lambert solver
I gave up trying to understand descriptions of the solution, and just
translated the algorithm from an existing implementation.
However, there seems to be a bug in the handling of hyperbolic
transfers - we take an arcoth of a value in (-1, 0), leading to a NaN.
I don't know whether that bug exists in the original or whether I
introduced it in my translation.